Separate object store from ro tests and create a bucket per test
